requirements.txt
fastapi
uvicorn
folium
requests
python-dotenv